feat(mocker): add `automocker` entry point
