+++ packages/mocker/package.json
@@ -32,6 +32,10 @@
       "types": "./dist/redirect.d.ts",
       "default": "./dist/redirect.js"
     },
+    "./automock": {
+      "types": "./dist/automock.d.ts",
+      "default": "./dist/automock.js"
+    },
     "./register": {
       "types": "./dist/register.d.ts",
       "default": "./dist/register.js"
