+++ packages/mocker/src/node/index.ts
@@ -1,6 +1,7 @@
 export { createManualModuleSource } from '../utils'
-export { automockModule, automockPlugin } from './automockPlugin'
+export { automockModule } from './automock'
 export type { AutomockPluginOptions } from './automockPlugin'
+export type { automockPlugin } from './automockPlugin'
 export { dynamicImportPlugin } from './dynamicImportPlugin'
 export { hoistMocks, hoistMocksPlugin } from './hoistMocksPlugin'
 export type { HoistMocksPluginOptions, HoistMocksResult } from './hoistMocksPlugin'
